project: set pagarme/moment as dependency
Given that moment doesn't consider timezones. It changes the moment
dependency to pagarme's moment version that uses _moment-timezone_
setting `America/Sao_Paulo` as the default timezone. Doing this the
tests broke because considering the timezone the date objects was
instantiated wrongly. This commit fixes that problem setting hours to
these date objects.